Fixed moblie screen styling
When the screen gets small enough it changes the menu to be a percent
size of the screen instead of a fixed width.  There are still small
problems when the viewport is super small but I do not think that will
be a problem on many devices, but I am going to see if I could resolve
this.